requirements.txt
# AraGovAssist - Requirements

# Core ML/NLP
sentence-transformers>=2.2.2
faiss-cpu>=1.8.0
transformers>=4.35.0
torch>=2.1.0

# Data Processing
numpy>=1.24.3
pandas>=2.0.3
scikit-learn>=1.3.0

# Utilities
tqdm>=4.66.1
python-dotenv>=1.0.0

# Jupyter (for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Web UI
streamlit>=1.28.0

# LLM Integration
google-generativeai>=0.3.0

# Translation
googletrans==4.0.0-rc1

# Hybrid Retrieval
rank-bm25>=0.2.2
